pyproject.toml
[tool.poetry]
name = "robynfr"
version = "0.1.0"
description = ""
authors = ["Anderson Rodrigues <eusouanderson@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
robyn = "^0.72.2"
SQLAlchemy = "^2.0.43"
python-dotenv = "^1.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
